third_party/WebKit/LayoutTests/accessibility/table-row-with-aria-role-expected.txt

1a	1b
2a	2b
Button inside table


Button inside table

This makes sure that a table with ARIA roles on the tbody or rows is not treated as a data table.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS list.role is 'AXRole: AXList'
PASS item1.role is 'AXRole: AXListItem'
PASS item2.role is 'AXRole: AXListItem'
PASS table.role is 'AXRole: AXTable'
PASS row1.role is 'AXRole: AXRow'
PASS row2.role is 'AXRole: AXRow'
PASS successfullyParsed is true

TEST COMPLETE
